% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GabesMyclt.R
\name{GabesMyclt}
\alias{GabesMyclt}
\title{Lab 8 function}
\usage{
GabesMyclt(n, iter)
}
\arguments{
\item{n}{A sample size}

\item{iter}{number of iterations description}
}
\value{
a histogram
}
\description{
Lab 8 function
}
\examples{
GabesMyclt(n=10, iter=10000)
}
