% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlikg2.R
\name{maxlikg2}
\alias{maxlikg2}
\title{My first function}
\usage{
maxlikg2(theta1, theta2, lfun = "logbinpois", ...)
}
\arguments{
\item{theta1}{A sample vector}

\item{theta2}{A sample vector}

\item{lfun}{a function to be applied to both inputed theta samples}
}
\value{
a 2 parameters with a general log  likelihood  estimator
}
\description{
My first function
}
\examples{
maxlikg2(theta1=seq(0,1,length=1000),theta2=seq(0,10,length=1000),nlevels=20)
}
