% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myboot2}
\alias{myboot2}
\title{My first function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations for sampling}

\item{x}{the sample size}

\item{fun}{the function like mean}

\item{alpha}{1 - alpha is the confidence % of the interval}

\item{cx}{variable that changes the character size}
}
\value{
a vector of squared components
}
\description{
My first function
}
\examples{
myfun(1:10)
}
