% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My confidence interval function}
\usage{
myci(x)
}
\arguments{
\item{x}{a sample to get ci for}
}
\value{
a confidence interval
}
\description{
My confidence interval function
}
\examples{
myci(x)
}
