% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{My first function}
\usage{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of tickets on a plane}

\item{gamma}{the probability that the plane will be truly overbooked}

\item{p}{probability for each ticket holder to show up to the flight}
}
\value{
a vector of squared components
}
\description{
My first function
}
\examples{
ntickets(N=400, gamma = 0.02, p = 0.95)
ntickets()
}
